requirements.txt
google-genai>=1.0.0
pytrends>=4.9.2
python-dotenv>=1.0.0
rich>=13.7.0
# ── Video Creator Agent ──────────────────────────────────────────────────
edge-tts>=6.1.9
moviepy==1.0.3
Pillow>=10.0.0
imageio-ffmpeg
requests>=2.31.0
